+++ tests/test_gpqr.py
@@ -4,7 +4,7 @@ from gpytorch.means import ConstantMean
 from gpytorch.mlls import VariationalELBO
 from gpytorch.variational import CholeskyVariationalDistribution, VariationalStrategy
 
-from gpytorch_qr.gpqr import ALDLikelihood, BatchQuantileGP
+from gpytorch_qr.gpqr import BatchALDLikelihood, BatchQuantileGP
 
 
 def test_gpqr():
@@ -41,7 +41,7 @@ def test_gpqr():
 
     inducing_points = torch.linspace(0, 1, 10).reshape(-1, 1)
     gp = MyQuantileGP(inducing_points, len(q))
-    likelihood = ALDLikelihood(q)
+    likelihood = BatchALDLikelihood(q)
 
     gp.train()
     likelihood.train()